introduce `signature` proc-macro
previously, to prevent implementing functions with wrong signatures, we
checked at comptime if the signatures of the main module and c matched
by hijacking `_` (type inference) and the fact that arrays can only
contain the same type and force the compiler to verify. now this is done
by a attribute macro that does the same thing, but in a less error prone
manner.